A matrix implementation for Koller's algorithm
This is a first attempt to implement Koller’s algorithm in a Matrix
fashion. We remove the stride method, since this can be computed on the
fly using matrices. We also add two more helpers to discover
assignments and indices in factors. This implementation is around 57%
faster than the current one in PGMPY, for big Factors (around 15 binary
variables), and around 97% faster for small Factors (around 2 binary
variables).