Removed Compatibility issues from mplp and its test
Through backtracking the error I found that there was issue with
decoded_result_assignment which gave hint that self.objective was having
some issue.After printing the state of self.objective in py2 and py3 i
realized that assignment of self.objective was somewhere wrong. Printed
the values of assignment in _update_message and found that updating was
wrong . After a some while realized that division in py2 doesn't give
float if numbers are integers. (see updation of phi in _update_message).