Fix joint distribution error during Exact Inference
This may be the same bug identified by @Erotemic in issue #534

I noticed that, when given multiple variables, _variable_elimination computes
them each independently, marginalizing over the others. This is a problem when
the joint distribution is not independent. I added a flag, marginal_independent,
which, if set to False, will return the joint factor for all variables queried.
It defaults to True in order to maintain backwards compatibility.

The default (old) behavior returns a dictionary where the keys are the variables
and the values are the factors. Since the new version only returns a single
joint probability distribution, it seems unnecessary to return it nested in a
dictionary. As a result, the new version simply returns the factor directly.
I realize this is a no-no as it means the method returns two different data
types based on which flag is set. But it carries the benefit that the query can
now be nested directly in anything that expects a Factor. I hope that this
improved ease of use justifies the breaking of convention, and that we can
eventually drop the old way entirely.