Fixes issue when a disconnected node was in the network file [fixes #936]:
The get_model method in all the Reader classes were creating the models using the edges in the network but when
a disconnected node was present in the network, it was failing because of not adding the node to the network.
This fix uses both add_nodes_from and add_edges_from to fix the issue with disconnected nodes.