Performance optimization for HillClimbSearch [fixes #939]:
Since hill climb search algorithm makes a lot of calls to BaseEstimator.state_counts
with the same arguments and computing them all makes them quite slow. So, this commit
adds a lru_cache on it, so that we don't have to compute the values everytime. The
default value of cache size is 2048 right now, but in future we would like to have
a dynamic size allocation based on system memory. Cache of this size is takes around
1.5 GB of memory and should have 100% hits (except for first computations) on learning
models of 12-15 variables.

This is a combination of 2 commits.
This is the 1st commit message:

adds lru cache

This is the commit message #2:

fix tests