Accelerate DAG.local_independencies method
Remove dfs function from DAG.local_independencies method and call
nx.dfs_preorder_nodes instead.

As dfs(nodes) doesn't remember which descendants of node were already
visited, the number of visits to each descendant is equal to the number
of paths between node and this descendant. As a result, the worst-case
time complexity is exponential in the number of nodes, instead of
quadratic for a classical depth-first-search traversal.