Renames default argument `bool=True` in chi_square function behavior.
This argument forced the function to check for a `significance_level`
argument upon return and changed the expected return value of the
chi_square function to be a boolean flag indicating if the found p_value was
within a particular significance level threshold. Changing the return
value of this function broke estimator classes that expected the
return value of this argument to be chi, dof, and p_value.

I don't think naming an argument after a data type is the most
intuitive, so I renamed this argument to `use_sig` and changed its
default value to False.

Resolves #1326