Bug fix: fix sampling bias bug caused by reusing random seed
Fix sampling bias bug from incorrectly reuing the seed for every node.

When passing to sample_discrete the same seed it is recycled for each node,
leading to sampling bias.
Now set the seed at the start of the sampling function, and do not pass
to sample_discrete() for each node.

This fixes the issue:
https://github.com/pgmpy/pgmpy/issues/1387