Added functionality to treesearch to do auto root and class selection
If root and/or class nodes are not provided when constructing the treesearch
algorithm, they are automatically picked as the node with the highest and
second highest sum of edge weights.

In practice this works really well to model any dataset, and is a really useful
feature when the dataset is large, and you don't know which features are important.

I added two corresponding unit tests for auto-selection for the chow-liu and tan methods.